refactor: model -> domain으로 변경
